--- client-react/src/pages/app/deployment-center/code/DeploymentCenterCodeCommandBar.tsx
@@ -1,216 +1,0 @@
-import React, { useContext } from 'react';
-import DeploymentCenterCommandBar from '../DeploymentCenterCommandBar';
-import { DeploymentCenterCodeCommandBarProps, WorkflowOption, SiteSourceControlRequestBody } from '../DeploymentCenter.types';
-import { GitHubCommit, GitHubActionWorkflowRequestContent } from '../../../../models/github';
-import { DeploymentCenterContext } from '../DeploymentCenterContext';
-import DeploymentCenterData from '../DeploymentCenter.data';
-import { BuildProvider, ScmType } from '../../../../models/site/config';
-import { getCodeAppWorkflowInformation } from '../utility/GitHubActionUtility';
-import { getArmToken, getWorkflowFilePath } from '../utility/DeploymentCenterUtility';
-import { PortalContext } from '../../../../PortalContext';
-import { useTranslation } from 'react-i18next';
-import { DeploymentCenterConstants } from '../DeploymentCenterConstants';
-import { getErrorMessage } from '../../../../ApiHelpers/ArmHelper';
-import { SiteStateContext } from '../../../../SiteState';
-import { DeploymentCenterPublishingContext } from '../DeploymentCenterPublishingContext';
-import LogService from '../../../../utils/LogService';
-import { LogCategories } from '../../../../utils/LogCategories';
-
-const DeploymentCenterCodeCommandBar: React.FC<DeploymentCenterCodeCommandBarProps> = props => {
-  const { isLoading, refresh, sync, formProps } = props;
-  const { t } = useTranslation();
-  const deploymentCenterContext = useContext(DeploymentCenterContext);
-  const deploymentCenterPublishingContext = useContext(DeploymentCenterPublishingContext);
-  const siteStateContext = useContext(SiteStateContext);
-  const portalContext = useContext(PortalContext);
-  const deploymentCenterData = new DeploymentCenterData();
-
-  const deployKudu = () => {
-    const payload: SiteSourceControlRequestBody = {
-      repoUrl: getRepoUrl(),
-      branch: formProps.values.branch || 'master',
-      isManualIntegration: isManualIntegration(),
-      isGitHubAction: formProps.values.buildProvider === BuildProvider.GitHubAction,
-      isMercurial: false,
-    };
-
-    if (formProps.values.sourceProvider === ScmType.LocalGit) {
-      return deploymentCenterData.patchSiteConfig(deploymentCenterContext.resourceId, {
-        properties: {
-          scmType: 'LocalGit',
-        },
-      });
-    } else {
-      return deploymentCenterData.updateSourceControlDetails(deploymentCenterContext.resourceId, { properties: payload });
-    }
-  };
-
-  const isManualIntegration = (): boolean => {
-    switch (formProps.values.sourceProvider) {
-      case ScmType.GitHub:
-      case ScmType.BitbucketGit:
-      case ScmType.LocalGit:
-        return false;
-      case ScmType.OneDrive:
-      case ScmType.Dropbox:
-      case ScmType.ExternalGit:
-        return true;
-      default:
-        LogService.error(
-          LogCategories.deploymentCenter,
-          'DeploymentCenterCodeCommandBar',
-          `Incorrect Source Provider ${formProps.values.sourceProvider}`
-        );
-        throw Error(`Incorrect Source Provider ${formProps.values.sourceProvider}`);
-    }
-  };
-
-  const getRepoUrl = (): string => {
-    switch (formProps.values.sourceProvider) {
-      case ScmType.GitHub:
-        return `${DeploymentCenterConstants.githubUri}/${formProps.values.org}/${formProps.values.repo}`;
-      case ScmType.BitbucketGit:
-        return `${DeploymentCenterConstants.bitbucketUrl}/${formProps.values.org}/${formProps.values.repo}`;
-      case ScmType.OneDrive:
-      case ScmType.Dropbox:
-        // TODO: (stpelleg): Pending Implementation of these ScmTypes
-        throw Error('Not implemented');
-      case ScmType.LocalGit:
-        //(note: stpelleg): Local Git does not require a Repo Url
-        return '';
-      case ScmType.ExternalGit:
-        if (formProps.values.externalUsername && formProps.values.externalPassword) {
-          const repoPath = formProps.values.repo.toLocaleLowerCase().replace('https://', '');
-          return `https://${formProps.values.externalUsername}:${formProps.values.externalPassword}@${repoPath}`;
-        }
-        return formProps.values.repo;
-      default:
-        LogService.error(
-          LogCategories.deploymentCenter,
-          'DeploymentCenterCodeCommandBar',
-          `Incorrect Source Provider ${formProps.values.sourceProvider}`
-        );
-        throw Error(`Incorrect Source Provider ${formProps.values.sourceProvider}`);
-    }
-  };
-
-  const deployGithubActions = async () => {
-    const repo = `${formProps.values.org}/${formProps.values.repo}`;
-    const branch = formProps.values.branch || 'master';
-
-    const workflowInformation = getCodeAppWorkflowInformation(
-      formProps.values.runtimeStack,
-      formProps.values.runtimeVersion,
-      formProps.values.runtimeRecommendedVersion,
-      branch,
-      siteStateContext.isLinuxApp,
-      formProps.values.gitHubPublishProfileSecretGuid,
-      deploymentCenterContext.siteDescriptor ? deploymentCenterContext.siteDescriptor.site : '',
-      deploymentCenterContext.siteDescriptor ? deploymentCenterContext.siteDescriptor.slot : ''
-    );
-
-    const commitInfo: GitHubCommit = {
-      repoName: repo,
-      branchName: branch,
-      filePath: getWorkflowFilePath(
-        branch,
-        deploymentCenterContext.siteDescriptor ? deploymentCenterContext.siteDescriptor.site : '',
-        deploymentCenterContext.siteDescriptor ? deploymentCenterContext.siteDescriptor.slot : ''
-      ),
-      message: t('githubActionWorkflowCommitMessage'),
-      contentBase64Encoded: btoa(workflowInformation.content),
-      committer: {
-        name: 'Azure App Service',
-        email: 'donotreply@microsoft.com',
-      },
-    };
-
-    const workflowConfigurationResponse = await deploymentCenterData.getWorkflowConfiguration(
-      formProps.values.org,
-      formProps.values.repo,
-      branch,
-      commitInfo.filePath,
-      deploymentCenterContext.gitHubToken
-    );
-
-    if (workflowConfigurationResponse.metadata.success) {
-      commitInfo.sha = workflowConfigurationResponse.data.sha;
-    }
-
-    const requestContent: GitHubActionWorkflowRequestContent = {
-      resourceId: deploymentCenterContext.resourceId,
-      secretName: workflowInformation.secretName,
-      commit: commitInfo,
-    };
-
-    return deploymentCenterData.createOrUpdateActionWorkflow(getArmToken(), deploymentCenterContext.gitHubToken, requestContent);
-  };
-
-  const deploy = async () => {
-    // NOTE(michinoy): Only initiate writing a workflow configuration file if the branch does not already have it OR
-    // the user opted to overwrite it.
-    if (
-      formProps.values.buildProvider === BuildProvider.GitHubAction &&
-      (formProps.values.workflowOption === WorkflowOption.Overwrite || formProps.values.workflowOption === WorkflowOption.Add)
-    ) {
-      const gitHubActionDeployResponse = await deployGithubActions();
-      if (!gitHubActionDeployResponse.metadata.success) {
-        return gitHubActionDeployResponse;
-      }
-    }
-
-    return deployKudu();
-  };
-
-  const saveGithubActionsDeploymentSettings = async () => {
-    const notificationId = portalContext.startNotification(t('settingupDeployment'), t('githubActionSavingSettings'));
-    const deployResponse = await deploy();
-    if (deployResponse.metadata.success) {
-      portalContext.stopNotification(notificationId, true, t('githubActionSettingsSavedSuccessfully'));
-    } else {
-      const errorMessage = getErrorMessage(deployResponse.metadata.error);
-      errorMessage
-        ? portalContext.stopNotification(notificationId, false, t('settingupDeploymentFailWithStatusMessage').format(errorMessage))
-        : portalContext.stopNotification(notificationId, false, t('settingupDeploymentFail'));
-    }
-  };
-
-  const saveAppServiceDeploymentSettings = async () => {
-    const notificationId = portalContext.startNotification(t('settingupDeployment'), t('settingupDeployment'));
-    const deployResponse = await deploy();
-    if (deployResponse.metadata.success) {
-      portalContext.stopNotification(notificationId, true, t('settingupDeploymentSuccess'));
-    } else {
-      const errorMessage = getErrorMessage(deployResponse.metadata.error);
-      errorMessage
-        ? portalContext.stopNotification(notificationId, false, t('settingupDeploymentFailWithStatusMessage').format(errorMessage))
-        : portalContext.stopNotification(notificationId, false, t('settingupDeploymentFail'));
-    }
-  };
-
-  const saveFunction = async () => {
-    if (formProps.values.buildProvider === BuildProvider.GitHubAction) {
-      await saveGithubActionsDeploymentSettings();
-    } else {
-      await saveAppServiceDeploymentSettings();
-    }
-    deploymentCenterContext.refresh();
-  };
-
-  const discardFunction = () => {
-    throw Error('not implemented');
-  };
-
-  return (
-    <DeploymentCenterCommandBar
-      saveFunction={saveFunction}
-      discardFunction={discardFunction}
-      showPublishProfilePanel={deploymentCenterPublishingContext.showPublishProfilePanel}
-      refresh={refresh}
-      isLoading={isLoading}
-      sync={sync}
-    />
-  );
-};
-
-export default DeploymentCenterCodeCommandBar;
